# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 블로그 자동화 시스템 API
  description: |
    블로그 자동화 시스템의 REST API 명세서입니다.
    Next.js App Router와 Supabase를 기반으로 구축되었습니다.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your-domain.com/api
    description: 프로덕션 서버
  - url: http://localhost:3000/api
    description: 개발 서버
security:
  - BearerAuth: []
paths:
  /posts:
    get:
      summary: 포스트 목록 조회
      description: 사용자의 포스트 목록을 페이지네이션과 함께 조회합니다.
      tags:
        - Posts
      parameters:
        - name: page
          in: query
          description: 페이지 번호
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 페이지당 항목 수
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: 성공적으로 포스트 목록을 조회했습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPostResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: 포스트 생성
      description: 새로운 포스트를 생성합니다.
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: 포스트가 성공적으로 생성되었습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: object
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0

    # 포스트 관련 스키마
    Post:
      type: object
      required:
        - id
        - title
        - content
        - status
        - channelId
        - workspaceId
        - authorId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
        status:
          type: string
          enum: [ideate, draft, image, seo, publish, published, failed, generating, ready]
        channelId:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            readTime:
              type: integer
              minimum: 0
            wordCount:
              type: integer
              minimum: 0
            tags:
              type: array
              items:
                type: string
            seoScore:
              type: integer
              minimum: 0
              maximum: 100
    CreatePostRequest:
      type: object
      required:
        - title
        - topic
        - channelIds
        - isPublic
        - tags
        - allowComments
        - notifyFollowers
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: 포스트 제목
        topic:
          type: string
          minLength: 1
          maxLength: 100
          description: 주제
        keywords:
          type: string
          maxLength: 500
          description: 키워드
        channelIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          description: 채널 ID 배열
        isPublic:
          type: boolean
          description: 공개 여부
        tags:
          type: array
          items:
            type: string
          description: 태그 배열
        scheduledAt:
          type: string
          format: date-time
          description: 예약 발행 시간
        allowComments:
          type: boolean
          description: 댓글 허용 여부
        notifyFollowers:
          type: boolean
          description: 팔로워 알림 여부
        content:
          type: string
          description: 포스트 내용
        excerpt:
          type: string
          maxLength: 300
          description: 요약
    PostResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Post'
    PaginatedPostResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Post'
            pagination:
              $ref: '#/components/schemas/Pagination'

    # 채널 관련 스키마
...
  responses:
    BadRequestError:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "입력 데이터가 올바르지 않습니다"
            code: "VALIDATION_ERROR"
    UnauthorizedError:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "인증이 필요합니다"
            code: "UNAUTHORIZED"
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "서버 오류가 발생했습니다"
            code: "INTERNAL_SERVER_ERROR"
...